Fix performance
